requirements.txt
anyio==3.6.2
bcrypt==4.0.1
boto3==1.26.78
botocore==1.29.78
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.92.0
fastapi-jwt-auth==0.5.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
mypy==1.0.0
mypy-extensions==1.0.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.4.0
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
types-pyasn1==0.4.0.4
types-python-jose==3.3.4.5
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
